requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.1

# HTTP Client
httpx==0.25.1
aiohttp==3.9.1

# Browser Automation
playwright==1.40.0

# Data Processing - Using versions with pre-built wheels
pandas==2.0.3
numpy==1.24.3
scipy==1.11.2
beautifulsoup4==4.12.2

# PDF Processing
pdfplumber==0.10.3
PyPDF2==3.0.1

# Image Processing
Pillow==10.1.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# LLM Integration
openai==1.3.7

# Geo-spatial Analysis - Optional, can be removed if causing issues
# geopandas==0.13.2
# shapely==2.0.1

# Network Analysis
networkx==3.1

# Machine Learning
scikit-learn==1.3.0

# Environment Management
python-dotenv==1.0.0

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
freezegun==1.4.0
